+++ cmd/collector/main.go
@@ -52,7 +52,7 @@ func getEnv(key, fallback string) string {
 	return fallback
 }
 
-const version = "1.2.107"
+const version = "1.2.108"
 
 type Collector struct {
 	cfg            *config.ProbeConfig